use generics
