+++ mol_gnn/featurizers/__init__.py
@@ -1,4 +1,5 @@
-from .atom import MultiHotAtomFeaturizer, AtomFeaturizer
-from .bond import MultiHotBondFeaturizer, BondFeaturizer
+from .atom import MultiHotAtomFeaturizer
+from .bond import MultiHotBondFeaturizer
 from .molgraph import *
 from .molecule import *
+from .base import Featurizer
