+++ mol_gnn/featurizers/molgraph/__init__.py
@@ -1,8 +1,4 @@
+from .base import MolGraphFeaturizer
 from .molgraph import MolGraph
-from .molecule import MoleculeMolGraphFeaturizer, BaseMoleculeMolGraphFeaturizer
-from .reaction import (
-    RxnMolGraphFeaturizer,
-    CondensedGraphOfReactionFeaturizer,
-    CGRFeaturizer,
-    RxnMode,
-)
+from .molecule import SimpleMoleculeMolGraphFeaturizer
+from .reaction import CondensedGraphOfReactionFeaturizer, CGRFeaturizer, RxnMode
