+++ mol_gnn/featurizers/molgraph/molgraph.py
@@ -13,4 +13,5 @@ class MolGraph(NamedTuple):
     edge_index: np.ndarray
     """an array of shape ``2 x E`` containing the edges of the graph in COO format"""
     rev_index: np.ndarray
-    """A array of shape ``E`` that maps from an edge index to the index of the source of the reverse edge in :attr:`edge_index` attribute."""
+    """A array of shape ``E`` that maps from an edge index to the index of the source of the
+    reverse edge in :attr:`edge_index` attribute."""
